main/requirements.txt
setproctitle==1.1.10
numpy==1.12.0
pyglib==0.1
scikit_image==0.9.3
tensorflow_gpu==1.1.0
python_gflags==3.1.1
python_magic==0.4.13

